Convenience functions for validators
